# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CarDealer CRUD",
        "description": "API documentation for the CarDealer CRUD application.",
        "license": {
            "name": "MIT License",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/cars": {
            "get": {
                "tags": [
                    "Cars"
                ],
                "summary": "Get list of cars",
                "operationId": "getCars",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "Cars"
                ],
                "summary": "Create a new car",
                "operationId": "storeCar",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/carsRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
        "/cars/{car}": {
            "get": {
                "tags": [
                    "Cars"
                ],
                "summary": "Show the car details",
                "operationId": "showCar",
                "parameters": [
                    {
                        "name": "car",
                        "in": "path",
                        "description": "ID of the car",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "put": {
                "tags": [
                    "Cars"
                ],
                "summary": "Update a car",
                "operationId": "updateCar",
                "parameters": [
                    {
                        "name": "car",
                        "in": "path",
                        "description": "ID of the car",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/carsRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Cars"
                ],
                "summary": "Delete a car",
                "operationId": "deleteCar",
                "parameters": [
                    {
                        "name": "car",
                        "in": "path",
                        "description": "ID of the car",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/cars/edit/{car}": {
            "get": {
                "tags": [
                    "Cars"
                ],
                "summary": "Show the car edit form",
                "operationId": "editCarForm",
                "parameters": [
                    {
                        "name": "car",
                        "in": "path",
                        "description": "ID of the car",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/search": {
            "post": {
                "tags": [
                    "Cars"
                ],
                "summary": "Search for cars",
                "operationId": "searchCars",
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SearchCarsRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "SearchCarsRequest": {
                "title": "Search Cars Request",
                "description": "Request schema for searching cars",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "model": {
                        "type": "string"
                    },
                    "fuelType": {
                        "type": "string"
                    },
                    "yearOfProduction": {
                        "type": "integer"
                    },
                    "transmission": {
                        "type": "string"
                    },
                    "driveType": {
                        "type": "string"
                    },
                    "price_min": {
                        "type": "number"
                    },
                    "price_max": {
                        "type": "number"
                    }
                },
                "type": "object"
            },
...
            "carsRequest": {
                "title": "Cars Request",
                "description": "Request body for creating or updating a car",
                "properties": {
                    "name": {
                        "description": "The name of the car",
                        "type": "string",
                        "example": "Toyota"
                    },
                    "model": {
                        "description": "The model of the car",
                        "type": "string",
                        "example": "Camry"
                    },
                    "fuelType": {
                        "description": "The type of fuel used by the car",
                        "type": "string",
                        "example": "Gasoline"
                    },
                    "yearOfProduction": {
                        "description": "The year of production of the car",
                        "type": "integer",
                        "example": 2022
                    },
                    "transmission": {
                        "description": "The type of transmission of the car",
                        "type": "string",
                        "example": "Automatic"
                    },
                    "driveType": {
                        "description": "The type of drive of the car",
                        "type": "string",
                        "example": "AWD"
                    },
                    "addInfo": {
                        "description": "Additional information about the car",
                        "type": "string",
                        "example": "Great condition, low mileage"
                    },
                    "image_path": {
                        "description": "Image of the car",
                        "type": "string",
                        "format": "binary"
                    },
                    "price": {
                        "description": "The price of the car",
                        "type": "number",
                        "format": "float",
                        "example": 25000.99
                    }
                },
                "type": "object"
            },
...
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ],
...
}